Add support for querying remote proxmox server for DNS.
This change allows proxmox-dns-server to run on a remote machine
and support DNS queries for multiple nodes. This way, we don't
interfere with proxmox's local dns cache and also query multiple
nodes in a cluster without running multiple instances. This can
allow us to use this as "true" nameserver for our purposes.